contracts/scripts/pyproject.toml
[project]
name = "test"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "eth-account>=0.13.7",
    "python-dotenv>=1.2.2",
    "web3>=7.14.1",
]
